Extract tight_binding parameters
